+++ tests/utils/test_utils.py
@@ -1,0 +1,18 @@
+from pydantic_resolve import util
+
+def test_get_class_field_annotations():
+    class C:
+        hello: str
+
+        def __init__(self, c: str):
+            self.c = c
+        
+    class D(C):
+        pass
+
+    class E(C):
+        world: str
+    
+    assert list(util.get_class_field_annotations(C)) == ['hello']
+    assert list(util.get_class_field_annotations(D)) == []
+    assert list(util.get_class_field_annotations(E)) == ['world']