+++ changelog.md
@@ -1,6 +1,12 @@
 # Changelog
 
+## v0.5.0 (2023.6.1)
+
+- `Resolver.ensure_type`: True will ensure `resolve_*` methods have return annotation.
+- add FastAPI integrated example `examples/fastapi_demo`
+
 ## v0.4.0 (2023.4.6)
+
 - add new install option `pip install "pydantic-resolve[dataloader]" to include `aiodataloader` by default
 - add new `doc/loader-cn.md`, `doc/loader-en.md` to explain the convinence of using `LoaderDepen`
 - add new params in Resolver: `loader_filters` to support global filter setting for inside loaders.